splite to standalone files
